+++ Game/Sound/SoundManager.cpp
@@ -88,7 +88,7 @@ namespace app
 		void SoundManager::Init()
 		{
 			/** 一括ロード */
-			for (int i = 0; i < enSoundList_Num; i++) {
+			for (uint8_t i = 0; i < enSoundList_Num; i++) {
 				std::string fullPath = SOUND_DIR + m_soundDefs[i].fileName + EXTENSION;
 				g_soundEngine->ResistWaveFileBank(i, fullPath.c_str());
 			}
